extract store to ts file
